Update the toolkit to 1.8.4
Update dijit version to 1.8.4 to be up to date with the latest 1.8
release. Thanks to bamboo, a lot of assets have been moved to their
right location (mainly stylesheets and images since Rails separates
all assets in their respective directory).

Update Dijit::VERSION as well. As said in in the commit on the main
repo (robin850/dojo-rails@eae0f91cd81d69dce5f18fe9b2cd6aa0863581f2)
we are waiting the first patch to update to 1.9.x or 2.0.

:sparkles: